+++ src/debug.c
@@ -38,6 +38,7 @@
 #include "threads_mngr.h"
 #include "io_threads.h"
 #include "sds.h"
+#include "module.h"
 
 #include <arpa/inet.h>
 #include <signal.h>
@@ -263,7 +264,7 @@ void xorObjectDigest(serverDb *db, robj *keyobj, unsigned char *digest, robj *o)
         ValkeyModuleDigest md = {{0}, {0}, keyobj, db->id};
         moduleValue *mv = o->ptr;
         moduleType *mt = mv->type;
-        moduleInitDigestContext(md);
+        moduleInitDigestContext(&md);
         if (mt->digest) {
             mt->digest(&md, mv->value);
             xorDigest(digest, md.x, sizeof(md.x));
