+++ src/functions.h
@@ -55,6 +55,7 @@
 typedef struct functionLibInfo functionLibInfo;
 
 /* ValkeyModule type aliases for scripting engine structs and types. */
+typedef struct ValkeyModule ValkeyModule;
 typedef ValkeyModuleScriptingEngineCtx engineCtx;
 typedef ValkeyModuleScriptingEngineFunctionCtx functionCtx;
 typedef ValkeyModuleScriptingEngineCompiledFunction compiledFunction;
