+++ src/script.c
@@ -31,6 +31,7 @@
 #include "script.h"
 #include "cluster.h"
 #include "cluster_slot_stats.h"
+#include "module.h"
 
 scriptFlag scripts_flags_def[] = {
     {.flag = SCRIPT_FLAG_NO_WRITES, .str = "no-writes"},
